:recycle: api: move builders into complementary-certification context
